requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
Django==3.0.8
django-bootstrap4==2.2.0
django-crispy-forms==1.9.2
django-jquery==3.1.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
psycopg2-binary==2.8.5
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.3.1

requests~=2.24.0

djangorestframework-jwt~=1.11.0
zeep~=3.4.0
lxml~=4.5.2
PyJWT~=1.7.1
jdatetime~=3.6.2